requirements.txt
zappa
requests
boto3
Flask
requests
simplejson
Pillow